split AuthComponent into several smaller components for clarity
